Clean up a little bit, rename and move App* files into components app folder
