+++ src/components/app/app.test.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render } from '@testing-library/react';
-import App from './App';
+import App from './app.component';
 
 test('renders learn react link', () => {
   const { getByText } = render(<App />);
